[LeetCode] Solve 2433/FindTheOriginalArrayOfPrefixXor
XOR is commutative, this challenge looks as if it should be
marked as easy - but leetcode marks it as medium one.
Solved by keeping cumulative XOR, i.e. the result of XORing
all the numbers decoded so far - the next one will be
decoded by XORing this cumulative XOR with its counterpart
from the "resulting" array.
if (A ^ B == C), then (A ^ C) must equal B.