[LeetCode] Solve 2391/MinimumAmountOfTimeToCollectGarbage
This performs pretty badly, but it does compute correct results.
Most probably calc() method that is currently called separately for each
garbage type could be rewritten so that it would handle all garbage
types, thus requiring just one pass instead of three separate passes.
But I'm satisfied that it works as is.

Beats 22.66% of other solutions.