[LeetCode] Solve 2706/BuyTwoChocolates
Fun trivia: the hint for this easy problem suggests sorting the list and
picking two first entries. This is a nice one-liner, but not optimal
enough [sorting won't be better than O(n*logn)].
A faster approach, implemented here, is to traverse the array and store
the lowest two entries. This is O(n).
Beats 93.98% of C++ solutions.