[LeetCode] Solve 1637/WidestVerticalArea
This is just a bad problem. The solution needs to plain ignore Y
coordinates, so half of the input data is useless.
Then the wording of the problem is just so strange, makes it difficult
to understand what this problem is about. According to the problem
statement, this task is about calculating the "widest vertical area"
of a rectangle which has fixed, infinite height.

In the end, the problem boils down to storing the x-coordinates,
sorting them and finding the largest gap between two consecutive sorted
points along the X axis.