Start solving Day21 part A / AoC2024
This version does not work for the last code of the example input.
The reason is that it currently generates one pre-defined safe path
through the numpad / dirpad.
This path is not necessarily optimal when another dirpad layer
is introduced. For the example input, 68 keypresses on the lowest layer
are necessary, when just 64 would be enough for the optimal path.

In order to fix this, the code needs to be updated so that *all* valid
paths for numpad/dirpad are generated; then an optimal one would be
selected.
It looks as if a simple heuristic for choosing "best" path is to
favor keys close to the "A" button. But trying out all paths eliminates
the need for any heuristic.