Solve 1092/Shortest Common Supersequence
Solved in Golang.
Runtime complexity is easy, O(l1*l2) where l1/l2 is the length of both strings.
Auxiliary space complexity is bit harder to assess; I'm pretty sure
that it depends on both length of strings (needed for the DP array);
but then it also seems to depend on the length of the LCS in an unclear way.
Lower bound is O(l1*l2*len(LCS)); it's probably higher with my implementation.